+++ tests/test_model_lock.py
@@ -120,7 +120,15 @@ async def test_ceiplock(
     central, mock_client, _ = central_client_factory_with_homegear_client
     lock: CustomDpIpLock = cast(CustomDpIpLock, get_prepared_custom_data_point(central, "VCU9724704", 1))
     assert lock.usage == DataPointUsage.CDP_PRIMARY
-    assert lock.service_method_names == ("lock", "open", "unlock")
+    assert lock.service_method_names == (
+        "get_schedule",
+        "get_structured_schedule",
+        "lock",
+        "open",
+        "set_schedule",
+        "set_structured_schedule",
+        "unlock",
+    )
 
     assert lock.is_locked is False
     await lock.lock()